+++ apps/demo/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @authhero/demo
 
+## 0.9.23
+
+### Patch Changes
+
+- Updated dependencies
+  - authhero@0.81.0
+  - @authhero/kysely-adapter@9.2.0
+
 ## 0.9.22
 
 ### Patch Changes